fix compiler warning about size_t to unsigned int conversion, changed Latch constructor to take unsigned int instead of uint_32_t to match std::atomic_uint

+++ src/vsg/io/read.cpp
@@ -95,7 +95,7 @@ PathObjects vsg::read(const Paths& filenames, ref_ptr<const Options> options)
         };
 
         // use latch to syncronize this thread with the file reading threads
-        ref_ptr<Latch> latch(new Latch(filenames.size()));
+        ref_ptr<Latch> latch(new Latch(static_cast<unsigned int>(filenames.size())));
 
         // add operations
         for (auto& [filename, object] : entries)
